# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTillEyes Marketplace Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  CLIENT: "CLIENT",
  USER: "USER",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: RegisterRequestRole;
  phone?: string;
  bio?: string;
  skills?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  ADMIN: "ADMIN",
  CLIENT: "CLIENT",
  USER: "USER",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  bio?: string | null;
  skills?: string | null;
  avatarUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  Open: "Open",
  Assigned: "Assigned",
  "In-Progress": "In-Progress",
  Submitted: "Submitted",
  Completed: "Completed",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  category?: string | null;
  budget?: number | null;
  deadline?: string | null;
  status: ProjectStatus;
  clientId: number;
  assignedUserId?: number | null;
  client?: User | null;
  assignedUser?: User | null;
  bidCount: number;
  createdAt: string;
  updatedAt: string;
}

export type BidStatus = (typeof BidStatus)[keyof typeof BidStatus];

export const BidStatus = {
  Pending: "Pending",
  Accepted: "Accepted",
  Rejected: "Rejected",
} as const;

export interface Bid {
  id: number;
  projectId: number;
  userId: number;
  amount: number;
  proposal?: string | null;
  timeline?: string | null;
  status: BidStatus;
  project?: Project | null;
  user?: User | null;
  createdAt: string;
}

export type SubmissionStatus =
  (typeof SubmissionStatus)[keyof typeof SubmissionStatus];

export const SubmissionStatus = {
  Pending: "Pending",
  Approved: "Approved",
  Forwarded: "Forwarded",
  Rejected: "Rejected",
} as const;

export interface Submission {
  id: number;
  projectId: number;
  userId: number;
  fileUrl: string;
  fileDescription?: string | null;
  notes?: string | null;
  status: SubmissionStatus;
  adminNotes?: string | null;
  project?: Project | null;
  user?: User | null;
  createdAt: string;
  updatedAt: string;
}

export type ProjectDetail = Project & {
  bids?: Bid[];
  submissions?: Submission[];
};

export interface CreateProjectRequest {
  title: string;
  description: string;
  category?: string;
  budget?: number;
  deadline?: string;
}

export type UpdateProjectRequestStatus =
  (typeof UpdateProjectRequestStatus)[keyof typeof UpdateProjectRequestStatus];

export const UpdateProjectRequestStatus = {
  Open: "Open",
  Assigned: "Assigned",
  "In-Progress": "In-Progress",
  Submitted: "Submitted",
  Completed: "Completed",
} as const;

export interface UpdateProjectRequest {
  title?: string;
  description?: string;
  category?: string;
  budget?: number;
  deadline?: string;
  status?: UpdateProjectRequestStatus;
}

export interface AssignProjectRequest {
  userId: number;
}

export interface CreateBidRequest {
  projectId: number;
  amount: number;
  proposal?: string;
  timeline?: string;
}

export interface CreateSubmissionRequest {
  projectId: number;
  fileUrl: string;
  fileDescription?: string;
  notes?: string;
}

export type UpdateSubmissionRequestStatus =
  (typeof UpdateSubmissionRequestStatus)[keyof typeof UpdateSubmissionRequestStatus];

export const UpdateSubmissionRequestStatus = {
  Pending: "Pending",
  Approved: "Approved",
  Forwarded: "Forwarded",
  Rejected: "Rejected",
} as const;

export interface UpdateSubmissionRequest {
  status?: UpdateSubmissionRequestStatus;
  adminNotes?: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface AdminDashboard {
  totalUsers: number;
  totalClients: number;
  totalWorkers: number;
  totalProjects: number;
  openProjects: number;
  assignedProjects: number;
  inProgressProjects: number;
  completedProjects: number;
  totalBids: number;
  pendingSubmissions: number;
  projectsByCategory: CategoryCount[];
  recentProjects: Project[];
}

export interface ClientDashboard {
  totalProjects: number;
  openProjects: number;
  inProgressProjects: number;
  completedProjects: number;
  totalBidsReceived: number;
  recentProjects: Project[];
  recentSubmissions: Submission[];
}

export interface UserDashboard {
  totalBids: number;
  acceptedBids: number;
  pendingBids: number;
  activeProjects: number;
  totalSubmissions: number;
  approvedSubmissions: number;
  availableProjects: number;
  recentBids: Bid[];
  assignedProjects: Project[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  project_created: "project_created",
  bid_submitted: "bid_submitted",
  project_assigned: "project_assigned",
  submission_uploaded: "submission_uploaded",
  project_completed: "project_completed",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  message: string;
  userName?: string | null;
  projectTitle?: string | null;
  createdAt: string;
}

export type GetUsersParams = {
  role?: GetUsersRole;
  search?: string;
};

export type GetUsersRole = (typeof GetUsersRole)[keyof typeof GetUsersRole];

export const GetUsersRole = {
  ADMIN: "ADMIN",
  CLIENT: "CLIENT",
  USER: "USER",
} as const;

export type GetProjectsParams = {
  status?: GetProjectsStatus;
  search?: string;
  category?: string;
};

export type GetProjectsStatus =
  (typeof GetProjectsStatus)[keyof typeof GetProjectsStatus];

export const GetProjectsStatus = {
  Open: "Open",
  Assigned: "Assigned",
  "In-Progress": "In-Progress",
  Submitted: "Submitted",
  Completed: "Completed",
} as const;

export type GetBidsParams = {
  projectId?: number;
  userId?: number;
};

export type GetSubmissionsParams = {
  projectId?: number;
  status?: GetSubmissionsStatus;
};

export type GetSubmissionsStatus =
  (typeof GetSubmissionsStatus)[keyof typeof GetSubmissionsStatus];

export const GetSubmissionsStatus = {
  Pending: "Pending",
  Approved: "Approved",
  Forwarded: "Forwarded",
  Rejected: "Rejected",
} as const;